Move everything under AllFutures module
and rename class to AllFutures::Base
