+++ lib/all_futures.rb
@@ -3,5 +3,5 @@
 require "active_entity/railtie"
 require "kredis"
 require "all_futures/version"
-require "all_futures/all_futures"
+require "all_futures/base"
 require "all_futures/callbacks"
